# docs/openapi.yml
openapi: 3.0.3
info:
  title: StreamFlix API
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /ratings/movie/{movieId}:
    post:
      summary: Añadir valoración a una película
      parameters:
        - in: path
          name: movieId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rating'
      responses:
        '200':
          description: Valoración añadida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rating'
    get:
      summary: Obtener valoraciones de una película
      parameters:
        - in: path
          name: movieId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de valoraciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rating'
  /ratings/user/{userId}:
    get:
      summary: Obtener valoraciones de un usuario
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de valoraciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rating'
  /ratings/{ratingId}:
    delete:
      summary: Eliminar valoración propia
      parameters:
        - in: path
          name: ratingId
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Valoración eliminada
        '403':
          description: No autorizado
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
    Movie:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        genre:
          type: string
        releaseYear:
          type: integer
        ageRating:
          type: string
    Rating:
      type: object
      properties:
        id:
          type: integer
        score:
          type: integer
        comment:
          type: string
        user:
          $ref: '#/components/schemas/User'
        movie:
          $ref: '#/components/schemas/Movie'